+++ RELEASE_NOTES.md
@@ -80,6 +80,7 @@ KeyboardKit 5.0 requires Xcode 13 and Swift 5.5.
 * `InputCalloutContext` `buttonFrame(for:)` has been removed.
 * `InputCalloutContext` `updateInput(for:geo:)` has been renamed to `updateInput(for:,in:)`.
 * `KeyboardBehavior` has a new `shouldSwitchToCapsLock` function.
+* `KeyboardCasing.neutral` has been removed.
 * `KeyboardInputSetProvider` functions have been converted to properties.
 * `KeyboardType.custom` has been renamed to `KeyboardType.custom(named:)`.
 * `SecondaryInputCalloutContext` init params have been reordered.